Fix for transactions grid collection not being reloaded properly
The extension relies on the fact that a new collection object is created
during the second call to _prepareCollection(). This doesn't happen with
the transactions grid, because it always first looks to see if there is
a collection already set. It does this because there is a subclass which
also creates a collection, and utilises this class to do most of the
work.

We don't need this for the main transactions grid, so we unset the
collection before calling _prepareCollection() for the 2nd time. This
makes everything work a lot more nicely.